move tx processing queue entry classes
